test: skip encrypted key tests in FIPS mode
Move passphrase/encrypted-key CertReloader tests to a !requirefips
build-tagged file since legacy PEM encryption uses MD5/DES which is
unavailable in FIPS mode.
